Move tags methods to Management and use them in the kotlinified code
