+++ src/main/java/org/cascadebot/cascadebot/commands/management/TagCreateSubCommand.java
@@ -24,7 +24,7 @@ public class TagCreateSubCommand implements ISubCommand {
         boolean warnUppercase = !context.getArg(0).equals(context.getArg(0).toLowerCase());
         String tagName = context.getArg(0).toLowerCase();
 
-        if (context.getData().getManagement().getTags().containsKey(tagName)) {
+        if (context.getData().getManagement().hasTag(tagName)) {
             context.getTypedMessaging().replyDanger(context.i18n("commands.tag.create.tag_already_exists", tagName));
             return;
         }
@@ -36,7 +36,7 @@ public class TagCreateSubCommand implements ISubCommand {
         }
 
         Tag tag = new Tag(context.getMessage(1), "tag", context.getArg(0));
-        context.getData().getCoreSettings().addTag(context.getArg(0), tag);
+        context.getData().getManagement().addTag(context.getArg(0), tag);
         context.getData().getPermissionsManager().registerGuildPermission(tag.getInternalPermission());
         context.getTypedMessaging().replySuccess(message);
     }
