+++ src/main/kotlin/org/cascadebot/cascadebot/data/objects/GuildSettingsManagement.kt
@@ -15,4 +15,20 @@ class GuildSettingsManagement {
     @Setting(directlyEditable = false)
     val permissions = GuildPermissions()
 
+    fun getTag(key: String): Tag? {
+        return tags[key]
+    }
+
+    fun hasTag(key: String): Boolean {
+        return tags.containsKey(key)
+    }
+
+    fun addTag(key: String, tag: Tag) {
+        tags[key] = tag
+    }
+
+    fun removeTag(key: String): Boolean {
+        return tags.remove(key) != null
+    }
+
 }
